requirements.txt
playwright
beautifulsoup4
pydantic
streamlit
pandas
google-generativeai
python-dotenv
openai
